// src/types/MovieType.ts
export type MovieType = {
    Poster: string;
    Genre: string;
    Ratings: {
        Source: string;
        Value: string;
    }[];
    Title: string;
    Language: string;
    Country: string;
    Plot: string;
    Awards: string;
    Director: string;
    Writer: string;
    Actors: string;
    imdbID: string;
    imdbRating: string;
    Released: string;
};